extract focus handling
